requirements.txt
streamlit
elevenlabs
python-dotenv
watchdog
datetime